host: Fix stopping RegisterHost.
When the connection terminates, Sampi needs to be notified, and close the stream
and unregister the host. We could probably further simplify the code if we
merged registerHost and RegisterHost.

We might consider sending a keep-alive heartbeat signal, or some other way of
ensuring the host is still up, as I do not know how reliable CloseNotifier is.

Signed-off-by: Blaž Hrastnik <blaz.hrast@gmail.com>